[framework] Moved public headers to a separate directory
To make physical structure match the framework layout
