+++ OAuth2Client/NXOAuth2ClientDelegate.h
