+++ OAuth2Client/NXOAuth2ConnectionDelegate.h
