+++ OAuth2Client/NXOAuth2Constants.h
