+++ OAuth2Client/NXOAuth2FileStreamWrapper.h
